requirements.txt
numpy==1.21.0
pandas>=1.3.0
requests==2.25.1
gitpython==3.1.18
transformers==4.28.1
